# api/openapi.yaml
openapi: 3.0.3
info:
  title: Ironhand API
  description: A secure vault service with end-to-end encryption.
  version: 1.0.0
servers:
  - url: /api/v1
    description: V1 API

components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: ironhand_session
      description: HTTP-only session cookie returned by /auth/register or /auth/login.

  schemas:
    CreateVaultRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

    CreateVaultResponse:
      type: object
      properties:
        vault_id:
          type: string
        member_id:
          type: string
        epoch:
          type: integer
          format: int64

    RegisterRequest:
      type: object
      required:
        - passphrase
      properties:
        passphrase:
          type: string

    RegisterResponse:
      type: object
      properties:
        secret_key:
          type: string

    LoginRequest:
      type: object
      required:
        - passphrase
        - secret_key
      properties:
        passphrase:
          type: string
        secret_key:
          type: string

    OpenVaultResponse:
      type: object
      properties:
        vault_id:
          type: string
        member_id:
          type: string
        epoch:
          type: integer
          format: int64

    VaultSummary:
      type: object
      properties:
        vault_id:
          type: string
        name:
          type: string
        description:
          type: string
        epoch:
          type: integer
          format: int64
        item_count:
          type: integer

    ListVaultsResponse:
      type: object
      properties:
        vaults:
          type: array
          items:
            $ref: '#/components/schemas/VaultSummary'

    ListItemsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: string

    PutItemRequest:
      type: object
      required:
        - data
      properties:
        data:
          type: string
          description: Base64 encoded encrypted data.
        content_type:
          type: string

    GetItemResponse:
      type: object
      properties:
        item_id:
          type: string
        data:
          type: string
          description: Base64 encoded encrypted data.

    UpdateItemRequest:
      type: object
      required:
        - data
      properties:
        data:
          type: string
          description: Base64 encoded encrypted data.
        content_type:
          type: string

    AddMemberRequest:
      type: object
      required:
        - member_id
        - pub_key
        - role
      properties:
        member_id:
          type: string
        pub_key:
          type: string
          description: Base64 encoded 32-byte public key.
        role:
          type: string
          enum: [owner, writer, reader]

    AddMemberResponse:
      type: object
      properties:
        epoch:
          type: integer
          format: int64

    ErrorResponse:
      type: object
      properties:
        error:
          type: string

paths:
  /auth/register:
    post:
      summary: Register a new account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'

  /auth/login:
    post:
      summary: Login with passphrase and secret key
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
        '401':
          description: Invalid credentials

  /auth/logout:
    post:
      summary: Logout current session
      operationId: logout
      responses:
        '200':
          description: Logged out

  /vaults:
    get:
      summary: List vaults visible to authenticated member
      operationId: listVaults
      security:
        - SessionAuth: []
      responses:
        '200':
          description: Vault list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListVaultsResponse'
        '401':
          description: Unauthorized
    post:
      summary: Create a new vault
      operationId: createVault
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVaultRequest'
      responses:
        '201':
          description: Vault created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateVaultResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /vaults/{vaultID}/open:
    post:
      summary: Open a vault
      operationId: openVault
      security:
        - SessionAuth: []
      parameters:
        - name: vaultID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vault opened successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenVaultResponse'
        '401':
          description: Unauthorized
        '404':
          description: Vault not found

  /vaults/{vaultID}/items:
    get:
      summary: List items in a vault
      operationId: listItems
      security:
        - SessionAuth: []
      parameters:
        - name: vaultID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of item IDs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListItemsResponse'
        '401':
          description: Unauthorized

  /vaults/{vaultID}/items/{itemID}:
    get:
      summary: Get an item from a vault
      operationId: getItem
      security:
        - SessionAuth: []
      parameters:
        - name: vaultID
          in: path
          required: true
          schema:
            type: string
        - name: itemID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetItemResponse'
        '401':
          description: Unauthorized
        '404':
          description: Item not found
    post:
      summary: Put an item into a vault
      operationId: putItem
      security:
        - SessionAuth: []
      parameters:
        - name: vaultID
          in: path
          required: true
          schema:
            type: string
        - name: itemID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutItemRequest'
      responses:
        '201':
          description: Item stored successfully
        '401':
          description: Unauthorized
    put:
      summary: Update an item in a vault
      operationId: updateItem
      security:
        - SessionAuth: []
      parameters:
        - name: vaultID
          in: path
          required: true
          schema:
            type: string
        - name: itemID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItemRequest'
      responses:
        '200':
          description: Item updated successfully
        '401':
          description: Unauthorized
        '404':
          description: Item not found
    delete:
      summary: Delete an item from a vault
      operationId: deleteItem
      security:
        - SessionAuth: []
      parameters:
        - name: vaultID
          in: path
          required: true
          schema:
            type: string
        - name: itemID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item deleted successfully
        '401':
          description: Unauthorized

  /vaults/{vaultID}/members:
    post:
      summary: Add a member to a vault
      operationId: addMember
      security:
        - SessionAuth: []
      parameters:
        - name: vaultID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
      responses:
        '201':
          description: Member added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddMemberResponse'
        '401':
          description: Unauthorized

  /vaults/{vaultID}/members/{memberID}:
    delete:
      summary: Revoke a member from a vault
      operationId: revokeMember
      security:
        - SessionAuth: []
      parameters:
        - name: vaultID
          in: path
          required: true
          schema:
            type: string
        - name: memberID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Member revoked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddMemberResponse'
        '401':
          description: Unauthorized

  /vaults/{vaultID}:
    delete:
      summary: Delete a vault
      operationId: deleteVault
      security:
        - SessionAuth: []
      parameters:
        - name: vaultID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vault deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden